--- .env.production
@@ -1,2 +1,0 @@
-MORALIS_NFTMARKETPLACE_APP_ID="L8kuo2v3DlsVmjGfmKNYWZq3W0ZQQSdADNav4hum"
-MORALIS_NFTMARKETPLACE_SERVER_URL="https://zz6umijjwcmt.usemoralis.com:2053/server"
