+++ src/components/Layout.js
@@ -1,5 +1,6 @@
 import Navbar from "./Navbar";
 
+
 const Layout = ({children}) => {
   return (
         <div className="content">
